fix(problem detail page): fixed PR review issues and add consts
#15
